# api.yml
openapi: 3.0.3
info:
  title: RSU/ESOP Platform API
  description: |
    API for managing equity compensation (RSUs/stock options) for startups.
    Features include pool management, grant administration, vesting calculations, and employee portal access.
    
    Key platform features:
    - Single ordinary-share class
    - Fixed DECIMAL(12,3) precision for all share & monetary amounts
    - 12-month cliff, 48-month total vesting
    - Roles: Admin, Employee
    - Pool & PPS management with vesting engine
    - Daily vesting batch job
    - Soft delete on all tables
  version: '1.0.0'
  contact:
    name: Platform Support
    email: support@rsu-platform.com
servers:
  - url: 'https://api.rsu-platform.com/api'
    description: Production API server
  - url: 'https://staging.rsu-platform.com/api'
    description: Staging server for testing
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0-issued JWT Access Token.
        Obtained via Auth0 Universal Login flow.
        Contains standard claims plus potential custom claims for role/permissions.
        Example claims expected by backend middleware:
        - sub: Auth0 User ID (used as auth0_user_id)
        - aud: API Identifier (audience)
        - iss: Issuer URL
        - exp: Expiration timestamp
        - iat: Issued at timestamp
        - scope: Permissions (if using RBAC)
        - Custom claims for role/tenant_id (if configured in Auth0 Rules/Actions)
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: 'INVALID_GRANT'
            message:
              type: string
              description: Human-readable error message
              example: 'The specified grant does not exist or is inaccessible'
            details:
              type: object
              description: Field-specific validation errors
              example:
                grant_id: 'Invalid UUID format'
                share_amount: 'Must be greater than 0.001'
                status: 'Must be one of: active, inactive'
    Success:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data, schema varies by endpoint
    Tenant:
      type: object
      required:
        - tenant_id
        - name
        - currency
        - timezone
        - created_at
      properties:
        tenant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          example: 'Acme Corp'
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: 'USD'
        timezone:
          type: string
          example: 'America/New_York'
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
          example: 'Acme Corporation'
        timezone:
          type: string
          example: 'America/Los_Angeles'
...
    EquityPool:
      type: object
      required:
        - pool_id
        - tenant_id
        - initial_amount
        - total_pool
        - created_at
      properties:
        pool_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        tenant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        initial_amount:
          $ref: '#/components/schemas/DecimalString'
          description: Initial amount of shares in the pool.
        total_pool:
          $ref: '#/components/schemas/DecimalString'
          description: Current total size of the pool (initial + events).
        available_shares:
          $ref: '#/components/schemas/DecimalString'
          description: Calculated available shares (TotalPool - Granted + Returned).
          readOnly: true
        granted_shares:
          $ref: '#/components/schemas/DecimalString'
          description: Calculated shares currently granted and active.
          readOnly: true
        returned_shares:
          $ref: '#/components/schemas/DecimalString'
          description: Calculated shares returned from terminated/cancelled grants.
          readOnly: true
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
    PoolEvent:
      type: object
      required:
        - event_id
        - pool_id
        - tenant_id
        - amount
        - event_type
        - effective_date
        - created_at
      properties:
        event_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        pool_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        tenant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        amount:
          $ref: '#/components/schemas/DecimalString'
          description: 'Signed amount: positive for top-up, negative for reduction'
        event_type:
          type: string
          enum: [initial, top_up, reduction]
          example: 'top_up'
          description: 'Event type, used for audit and reporting'
        effective_date:
          type: string
          format: date
          example: '2025-05-01'
          description: 'Date when the pool change takes effect'
        notes:
          type: string
          example: 'Additional shares approved by board'
          description: 'Optional notes about the event'
        created_by:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
          description: 'User who created the event'
        created_at:
          type: string
          format: date-time
          example: '2025-05-01T10:00:00Z'
    PoolEventCreate:
      type: object
      required:
        - amount
        - event_type
        - effective_date
      properties:
        amount:
          $ref: '#/components/schemas/DecimalString'
          description: 'Signed amount: positive for top-up, negative for reduction. No CHECK constraint enforced.'
        event_type:
          type: string
          enum: [top_up, reduction]
          example: 'top_up'
        effective_date:
          type: string
          format: date
          example: '2025-05-01'
        notes:
          type: string
          example: 'Additional shares approved by board'
...
    PaginatedResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            type: object
        pagination:
          type: object
          required:
            - total_items
            - total_pages
            - current_page
            - limit
          properties:
            total_items:
              type: integer
              example: 45
            total_pages:
              type: integer
              example: 3
            current_page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            next_page:
              type: integer
              nullable: true
              example: 2
            prev_page:
              type: integer
              nullable: true
              example: null
...
    DecimalString:
      type: string
      description: String representation of a DECIMAL(12,3) value.
      pattern: '^-?[0-9]{1,9}(\\.[0-9]{1,3})?$'
      example: '123456.789'
...
  parameters:
...
    UserIdParam:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the user
...
    PoolIdParam:
      name: pool_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the equity pool
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page for pagination
...
    FromDateParam:
      name: from_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter by date range (start date)
    ToDateParam:
      name: to_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter by date range (end date)
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /tenant:
    get:
      summary: Get current tenant details
      description: Retrieve current tenant information based on authenticated user's token.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tenant details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    
    patch:
      summary: Update current tenant
      description: Update current tenant information (e.g., name, timezone). Requires Admin role.
      security:
        - bearerAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdate'
      responses:
        '200':
          description: Tenant updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    
    delete:
      summary: Delete user
      description: Soft delete a user (sets deleted_at timestamp)
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /pools:
    get:
      summary: Get equity pool
      description: Retrieve equity pool information for the tenant
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pool retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EquityPool'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /pools/{pool_id}/events:
    get:
      summary: List pool events
      description: Retrieve paginated list of pool events
      parameters:
        - $ref: '#/components/parameters/PoolIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/FromDateParam'
        - $ref: '#/components/parameters/ToDateParam'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pool events retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginatedResponse'
                          - type: object
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/PoolEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    
    post:
      summary: Create pool event
      description: Add a new event to the equity pool
      parameters:
        - $ref: '#/components/parameters/PoolIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolEventCreate'
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Pool event created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PoolEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...